requirements.txt
pandas
markdown
mysql-connector-python
